biblioteca para a decomposta LU
