+++ trabalho1/decompostaLU.h
@@ -1,0 +1,3 @@
+#define MAX 100
+
+bool decomposicaoLU(int n, double a[][MAX], double b[], double x[])
